% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{banana_gradloglikelihood}
\alias{banana_gradloglikelihood}
\title{Evaluate gradient of banana example loglikelihood function}
\usage{
banana_gradloglikelihood(x, like_alpha, like_beta)
}
\arguments{
\item{x}{evaluation points}
}
\value{
gradient values
}
\description{
Evaluate gradient of banana example loglikelihood function
}
