% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{banana_gradlogprior}
\alias{banana_gradlogprior}
\title{Evaluate gradient of banana example prior density}
\usage{
banana_gradlogprior(x)
}
\arguments{
\item{x}{evaluation points}
}
\value{
gradient values
}
\description{
Evaluate gradient of banana example prior density
}
