% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{banana_loglikelihood}
\alias{banana_loglikelihood}
\title{Evaluate banana example loglikelihood function}
\usage{
banana_loglikelihood(x, like_alpha, like_beta)
}
\arguments{
\item{x}{evaluation points}
}
\value{
density values
}
\description{
Evaluate banana example loglikelihood function
}
