% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{banana_logprior}
\alias{banana_logprior}
\title{Evaluate banana example prior density}
\usage{
banana_logprior(x)
}
\arguments{
\item{x}{evaluation points}
}
\value{
density values
}
\description{
Evaluate banana example prior density
}
