% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{banana_partialderivative_loglikelihood}
\alias{banana_partialderivative_loglikelihood}
\title{Evaluate partial derivative of banana example likelihood function}
\usage{
banana_partialderivative_loglikelihood(dim, x, like_alpha, like_beta)
}
\arguments{
\item{dim}{dimension along which partial derivative is taken (dim = 0 is the first coordinate)}

\item{x}{evaluation point}
}
\value{
partial derivative value
}
\description{
Evaluate partial derivative of banana example likelihood function
}
