% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{banana_sampleprior}
\alias{banana_sampleprior}
\title{Sample from banana example prior distribution}
\usage{
banana_sampleprior(n)
}
\arguments{
\item{n}{number of samples}
}
\value{
samples
}
\description{
Sample from banana example prior distribution
}
