% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{baseball_artificial_logprior}
\alias{baseball_artificial_logprior}
\title{Evaluate variance components model artificial prior density on baseball dataset}
\usage{
baseball_artificial_logprior(x)
}
\arguments{
\item{x}{evaluation points}
}
\value{
density values
}
\description{
Evaluate variance components model artificial prior density on baseball dataset
}
