% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{baseball_gibbsflow}
\alias{baseball_gibbsflow}
\title{Compute Gibbs flow for variance components model on baseball dataset}
\usage{
baseball_gibbsflow(
  stepsize,
  lambda,
  lambda_next,
  derivative_lambda,
  xparticles,
  logdensity
)
}
\arguments{
\item{stepsize}{numerical integration step size}

\item{lambda}{tempering level}

\item{lambda_next}{next tempering level}

\item{derivative_lambda}{time derivative of tempering schedule}

\item{xparticles}{particle locations}

\item{logdensity}{corresponding log density values}
}
\value{
list with keys:
\code{xparticles} new particle locations,
\code{log_jacobian_dets} log Jacobian determinant of the flow
}
\description{
Compute Gibbs flow for variance components model on baseball dataset
}
