% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{baseball_gibbsvelocity}
\alias{baseball_gibbsvelocity}
\title{Compute Gibbs velocity for variance components model on baseball dataset}
\usage{
baseball_gibbsvelocity(time, xparticles, exponent)
}
\arguments{
\item{time}{time}

\item{xparticles}{particle positions}

\item{exponent}{exponent of tempering schedule}
}
\value{
gibbs_velocity gibbs velocity field
}
\description{
Compute Gibbs velocity for variance components model on baseball dataset
}
