% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{baseball_gradloglikelihood}
\alias{baseball_gradloglikelihood}
\title{Evaluate gradient of variance components loglikelihood on baseball dataset}
\usage{
baseball_gradloglikelihood(x)
}
\arguments{
\item{x}{evaluation points}
}
\value{
gradient values
}
\description{
Evaluate gradient of variance components loglikelihood on baseball dataset
}
