% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{baseball_gradlogprior}
\alias{baseball_gradlogprior}
\title{Evaluate gradient of variance components prior density on baseball dataset}
\usage{
baseball_gradlogprior(x)
}
\arguments{
\item{x}{evaluation points}
}
\value{
gradient values
}
\description{
Evaluate gradient of variance components prior density on baseball dataset
}
