% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{baseball_loglikelihood}
\alias{baseball_loglikelihood}
\title{Evaluate variance components model loglikelihood function on baseball dataset}
\usage{
baseball_loglikelihood(x)
}
\arguments{
\item{x}{evaluation points}
}
\value{
density values
}
\description{
Evaluate variance components model loglikelihood function on baseball dataset
}
