% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{baseball_sample_artificial_prior}
\alias{baseball_sample_artificial_prior}
\title{Sample from variance components model artificial prior distribution}
\usage{
baseball_sample_artificial_prior(N)
}
\arguments{
\item{n}{number of samples}
}
\value{
samples
}
\description{
Sample from variance components model artificial prior distribution
}
