% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_kernel.R
\name{construct_kernel}
\alias{construct_kernel}
\title{Constructs MCMC kernel}
\usage{
construct_kernel(logtarget, gradlogtarget, mcmc)
}
\arguments{
\item{logtarget}{evaluates log density of the target distribution}

\item{gradlogtarget}{evaluates gradient of the log target density}

\item{mcmc}{list with keys: 
\code{choice} specifies type of MCMC method, 
\code{parameters} specifies algorithmic tuning parameters,}
}
\value{
function which returns desired MCMC kernel
}
\description{
Constructs MCMC kernel
}
