% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{coxprocess_gibbsflow_vi_proposal}
\alias{coxprocess_gibbsflow_vi_proposal}
\title{Compute Gibbs flow for Cox process model with VI Gaussian proposal initialization}
\usage{
coxprocess_gibbsflow_vi_proposal(
  stepsize,
  lambda,
  derivative_lambda,
  xparticles,
  logdensity,
  counts
)
}
\arguments{
\item{stepsize}{numerical integration step size}

\item{lambda}{tempering level}

\item{derivative_lambda}{time derivative of tempering schedule}

\item{xparticles}{particle locations}

\item{logdensity}{corresponding log density values}

\item{counts}{dataset}
}
\value{
list with keys:
\code{xparticles} new particle locations,
\code{log_jacobian_dets} log Jacobian determinant of the flow
}
\description{
Compute Gibbs flow for Cox process model with VI Gaussian proposal initialization
}
