% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{coxprocess_gibbsvelocity_vi_proposal}
\alias{coxprocess_gibbsvelocity_vi_proposal}
\title{Compute Gibbs velocity for Cox process model with VI Gaussian proposal initialization}
\usage{
coxprocess_gibbsvelocity_vi_proposal(time, xparticles, exponent, counts)
}
\arguments{
\item{time}{time}

\item{xparticles}{particle positions}

\item{exponent}{exponent of tempering schedule}

\item{counts}{dataset}
}
\value{
gibbs_velocity gibbs velocity field
}
\description{
Compute Gibbs velocity for Cox process model with VI Gaussian proposal initialization
}
