% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{coxprocess_gradlog_vi_proposal}
\alias{coxprocess_gradlog_vi_proposal}
\title{Evaluate gradient of VI Gaussian proposal density}
\usage{
coxprocess_gradlog_vi_proposal(x)
}
\arguments{
\item{x}{evaluation points}
}
\value{
gradient values
}
\description{
Evaluate gradient of VI Gaussian proposal density
}
