% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{coxprocess_gradloglikelihood}
\alias{coxprocess_gradloglikelihood}
\title{Evaluate gradient of Cox process loglikelihood function}
\usage{
coxprocess_gradloglikelihood(x, counts)
}
\arguments{
\item{x}{evaluation points}
}
\value{
gradient values
}
\description{
Evaluate gradient of Cox process loglikelihood function
}
