% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{coxprocess_log_ep_proposal}
\alias{coxprocess_log_ep_proposal}
\title{Evaluate EP Gaussian proposal density}
\usage{
coxprocess_log_ep_proposal(x)
}
\arguments{
\item{x}{evaluation points}
}
\value{
density values
}
\description{
Evaluate EP Gaussian proposal density
}
