% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{coxprocess_log_ep_proposal_conditional}
\alias{coxprocess_log_ep_proposal_conditional}
\title{Evaluate conditional density of EP Gaussian proposal}
\usage{
coxprocess_log_ep_proposal_conditional(dim, x, other_x, other_index)
}
\arguments{
\item{dim}{dimension of interest (dim = 0 is the first coordinate)}

\item{x}{evaluation points along dim}

\item{other_x}{conditional values}

\item{other_index}{logical vector indexing conditioned coordinates}
}
\value{
density values
}
\description{
Evaluate conditional density of EP Gaussian proposal
}
