% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{coxprocess_log_vi_proposal}
\alias{coxprocess_log_vi_proposal}
\title{Evaluate VI Gaussian proposal density}
\usage{
coxprocess_log_vi_proposal(x)
}
\arguments{
\item{x}{evaluation points}
}
\value{
density values
}
\description{
Evaluate VI Gaussian proposal density
}
