% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{coxprocess_loglikelihood}
\alias{coxprocess_loglikelihood}
\title{Evaluate Cox process loglikelihood function}
\usage{
coxprocess_loglikelihood(x, counts)
}
\arguments{
\item{x}{evaluation points}
}
\value{
density values
}
\description{
Evaluate Cox process loglikelihood function
}
