% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{coxprocess_loglikelihood_term}
\alias{coxprocess_loglikelihood_term}
\title{Evaluate a particular term of Cox process loglikelihood function}
\usage{
coxprocess_loglikelihood_term(dim, x, counts)
}
\arguments{
\item{dim}{loglikelihood term depends only on this dimension (dim = 0 is the first coordinate)}

\item{x}{evaluation points}
}
\value{
density values
}
\description{
Evaluate a particular term of Cox process loglikelihood function
}
