% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{coxprocess_logprior}
\alias{coxprocess_logprior}
\title{Evaluate multivariate Gaussian prior density}
\usage{
coxprocess_logprior(x)
}
\arguments{
\item{x}{evaluation points}
}
\value{
density values
}
\description{
Evaluate multivariate Gaussian prior density
}
