% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{coxprocess_logprior_conditional}
\alias{coxprocess_logprior_conditional}
\title{Evaluate conditional density of multivariate Gaussian prior}
\usage{
coxprocess_logprior_conditional(dim, x, other_x)
}
\arguments{
\item{dim}{dimension of interest (dim = 0 is the first coordinate)}

\item{x}{evaluation points along dim}

\item{other_x}{conditional values}
}
\value{
density values
}
\description{
Evaluate conditional density of multivariate Gaussian prior
}
