% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{coxprocess_partialderivative_loglikelihood}
\alias{coxprocess_partialderivative_loglikelihood}
\title{Evaluate partial derivative of Cox process likelihood function}
\usage{
coxprocess_partialderivative_loglikelihood(dim, x, counts)
}
\arguments{
\item{dim}{dimension along which partial derivative is taken (dim = 0 is the first coordinate)}

\item{x}{evaluation point}
}
\value{
partial derivative value
}
\description{
Evaluate partial derivative of Cox process likelihood function
}
