% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{coxprocess_partialderivative_logprior}
\alias{coxprocess_partialderivative_logprior}
\title{Evaluate partial derivative of multivariate Gaussian prior density}
\usage{
coxprocess_partialderivative_logprior(dim, x)
}
\arguments{
\item{dim}{dimension along which partial derivative is taken (dim = 0 is the first coordinate)}

\item{x}{evaluation point}
}
\value{
partial derivative value
}
\description{
Evaluate partial derivative of multivariate Gaussian prior density
}
