% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{coxprocess_sample_vi_proposal}
\alias{coxprocess_sample_vi_proposal}
\title{Sample from VI Gaussian proposal distribution}
\usage{
coxprocess_sample_vi_proposal(n)
}
\arguments{
\item{n}{number of samples}
}
\value{
samples
}
\description{
Sample from VI Gaussian proposal distribution
}
