% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{coxprocess_sampleprior}
\alias{coxprocess_sampleprior}
\title{Sample from multivariate Gaussian prior distribution}
\usage{
coxprocess_sampleprior(n)
}
\arguments{
\item{n}{number of samples}
}
\value{
samples
}
\description{
Sample from multivariate Gaussian prior distribution
}
