% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{coxprocess_stein_variational_importance_sampling}
\alias{coxprocess_stein_variational_importance_sampling}
\title{Perform Stein variational importance sampling}
\usage{
coxprocess_stein_variational_importance_sampling(
  nfparticles,
  nlparticles,
  niterations,
  stepsize,
  counts
)
}
\arguments{
\item{nfparticles}{number of follower particles}

\item{nlparticles}{number of leader particles}

\item{niterations}{number of time iterations}

\item{stepsize}{stepsize of each iteration}

\item{counts}{dataset}
}
\value{
list with keys:
\code{xparticles} new particle locations
}
\description{
Perform Stein variational importance sampling
}
