% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gaussian_gibbsvelocity}
\alias{gaussian_gibbsvelocity}
\title{Compute Gibbs velocity for Gaussian model}
\usage{
gaussian_gibbsvelocity(time, xpoint, exponent)
}
\arguments{
\item{time}{time}

\item{xpoint}{particle position}

\item{exponent}{exponent of tempering schedule}
}
\value{
gibbs_velocity gibbs velocity field
}
\description{
Compute Gibbs velocity for Gaussian model
}
