% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gaussian_gradloglikelihood}
\alias{gaussian_gradloglikelihood}
\title{Evaluate gradient of multivariate Gaussian loglikelihood function}
\usage{
gaussian_gradloglikelihood(x)
}
\arguments{
\item{x}{evaluation points}
}
\value{
gradient values
}
\description{
Evaluate gradient of multivariate Gaussian loglikelihood function
}
