% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gaussian_gradlogprior}
\alias{gaussian_gradlogprior}
\title{Evaluate gradient of multivariate Gaussian prior density}
\usage{
gaussian_gradlogprior(x)
}
\arguments{
\item{x}{evaluation points}
}
\value{
gradient values
}
\description{
Evaluate gradient of multivariate Gaussian prior density
}
