% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gaussian_log_normconst}
\alias{gaussian_log_normconst}
\title{Compute log normalizing constant of multivariate Gaussian example}
\usage{
gaussian_log_normconst(lambda)
}
\arguments{
\item{lambda}{temperature}
}
\value{
log normalizing constant
}
\description{
Compute log normalizing constant of multivariate Gaussian example
}
