% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gaussian_loglikelihood}
\alias{gaussian_loglikelihood}
\title{Evaluate multivariate Gaussian loglikelihood function}
\usage{
gaussian_loglikelihood(x)
}
\arguments{
\item{x}{evaluation points}
}
\value{
density values
}
\description{
Evaluate multivariate Gaussian loglikelihood function
}
