% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gaussian_logposterior}
\alias{gaussian_logposterior}
\title{Evaluate multivariate Gaussian posterior density}
\usage{
gaussian_logposterior(x, lambda)
}
\arguments{
\item{x}{evaluation points}

\item{lambda}{temperature}
}
\value{
density values
}
\description{
Evaluate multivariate Gaussian posterior density
}
