% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gaussian_partialderivative_loglikelihood}
\alias{gaussian_partialderivative_loglikelihood}
\title{Evaluate partial derivative of multivariate Gaussian likelihood function}
\usage{
gaussian_partialderivative_loglikelihood(dim, x)
}
\arguments{
\item{dim}{dimension along which partial derivative is taken (dim = 0 is the first coordinate)}

\item{x}{evaluation point}
}
\value{
partial derivative value
}
\description{
Evaluate partial derivative of multivariate Gaussian likelihood function
}
