% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gaussian_posterior_cov}
\alias{gaussian_posterior_cov}
\title{Compute posterior covariance of multivariate Gaussian example}
\usage{
gaussian_posterior_cov(lambda)
}
\arguments{
\item{lambda}{temperature}
}
\value{
posterior covariance
}
\description{
Compute posterior covariance of multivariate Gaussian example
}
