% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gaussian_posterior_mean}
\alias{gaussian_posterior_mean}
\title{Compute posterior mean of multivariate Gaussian example}
\usage{
gaussian_posterior_mean(lambda)
}
\arguments{
\item{lambda}{temperature}
}
\value{
posterior mean
}
\description{
Compute posterior mean of multivariate Gaussian example
}
