% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gaussian_stein_variational_importance_sampling}
\alias{gaussian_stein_variational_importance_sampling}
\title{Perform Stein variational importance sampling}
\usage{
gaussian_stein_variational_importance_sampling(
  nfparticles,
  nlparticles,
  niterations,
  stepsize,
  bandwidth
)
}
\arguments{
\item{nfparticles}{number of follower particles}

\item{nlparticles}{number of leader particles}

\item{niterations}{number of time iterations}

\item{stepsize}{stepsize of each iteration}
}
\value{
list with keys:
\code{xparticles} new particle locations
\code{ess} effective sample size
\code{log_normconst} log normalizing constant estimate
}
\description{
Perform Stein variational importance sampling
}
