% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mala_kernel.R
\name{get_mala_kernel}
\alias{get_mala_kernel}
\title{Construct Metropolis adjusted Langevin algorithm kernel}
\usage{
get_mala_kernel(logtarget, gradlogtarget, parameters)
}
\arguments{
\item{logtarget}{evaluates log density of the target distribution}

\item{gradlogtarget}{evaluates gradient of the log target density}

\item{parameters}{list with key: 
\code{stepsize} specifies stepsize of discretization}
}
\value{
list with key: \code{kernel}
}
\description{
Construct Metropolis adjusted Langevin algorithm kernel
}
