% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwmh_kernel.R
\name{get_rwmh_kernel}
\alias{get_rwmh_kernel}
\title{Construct random walk Metropolis-Hastings kernel}
\usage{
get_rwmh_kernel(logtarget, parameters)
}
\arguments{
\item{logtarget}{evaluates log density of the target distribution}

\item{parameters}{list with key: 
\code{stepsize} specifies stepsize of random walk}
}
\value{
list with key: \code{kernel}
}
\description{
Construct random walk Metropolis-Hastings kernel
}
