\name{GibbsFlow-package}
\alias{GibbsFlow-package}
\alias{GibbsFlow}
\docType{package}
\title{
\packageTitle{GibbsFlow}
}
\description{
\packageDescription{GibbsFlow}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{GibbsFlow}
\packageIndices{GibbsFlow}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{GibbsFlow}

Maintainer: \packageMaintainer{GibbsFlow}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
