% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GibbsFlow.R
\docType{package}
\name{GibbsFlow}
\alias{GibbsFlow}
\title{GibbsFlow package}
\description{
Gibbs Flow for Approximate Transport
}
\details{
This package implements sequential Monte Carlo samplers that use Gibbs flow transport.
}
\author{
Jeremy Heng <jeremyhengjm@gmail.com>
}
\keyword{package}
