% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gradmvnpdf}
\alias{gradmvnpdf}
\title{Evaluate gradient of multivariate Gaussian prior density}
\usage{
gradmvnpdf(x, mu, precision)
}
\arguments{
\item{x}{evaluation points}

\item{mu}{mean vector}

\item{precision}{precision matrix}
}
\value{
gradient values
}
\description{
Evaluate gradient of multivariate Gaussian prior density
}
