% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mixtureexample_gibbsflow}
\alias{mixtureexample_gibbsflow}
\title{Compute Gibbs flow for mixture example}
\usage{
mixtureexample_gibbsflow(
  stepsize,
  lambda,
  derivative_lambda,
  xparticles,
  logdensity
)
}
\arguments{
\item{stepsize}{numerical integration step size}

\item{lambda}{tempering level}

\item{derivative_lambda}{time derivative of tempering schedule}

\item{xparticles}{particle locations}

\item{logdensity}{corresponding log density values}
}
\value{
list with keys:
\code{xparticles} new particle locations,
\code{log_jacobian_dets} log Jacobian determinant of the flow
}
\description{
Compute Gibbs flow for mixture example
}
