% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mixtureexample_gibbsvelocity}
\alias{mixtureexample_gibbsvelocity}
\title{Compute Gibbs flow for mixture example}
\usage{
mixtureexample_gibbsvelocity(time, xpoint, exponent)
}
\arguments{
\item{time}{time}

\item{xpoint}{particle position}

\item{exponent}{exponent of tempering schedule}
}
\value{
gibbs_velocity gibbs velocity field
}
\description{
Compute Gibbs flow for mixture example
}
