% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mixtureexample_gradloglikelihood}
\alias{mixtureexample_gradloglikelihood}
\title{Evaluate gradient of mixture example loglikelihood function}
\usage{
mixtureexample_gradloglikelihood(x)
}
\arguments{
\item{x}{evaluation points}
}
\value{
gradient values
}
\description{
Evaluate gradient of mixture example loglikelihood function
}
