% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mixtureexample_gradlogprior}
\alias{mixtureexample_gradlogprior}
\title{Evaluate gradient of mixture example prior density}
\usage{
mixtureexample_gradlogprior(x)
}
\arguments{
\item{x}{evaluation points}
}
\value{
gradient values
}
\description{
Evaluate gradient of mixture example prior density
}
