% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mixtureexample_log_normconst}
\alias{mixtureexample_log_normconst}
\title{Compute log normalizing constant of mixture example}
\usage{
mixtureexample_log_normconst()
}
\value{
log normalizing constant
}
\description{
Compute log normalizing constant of mixture example
}
