% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mixtureexample_loglikelihood}
\alias{mixtureexample_loglikelihood}
\title{Evaluate mixture example loglikelihood function}
\usage{
mixtureexample_loglikelihood(x)
}
\arguments{
\item{x}{evaluation points}
}
\value{
density values
}
\description{
Evaluate mixture example loglikelihood function
}
