% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mixtureexample_logposterior}
\alias{mixtureexample_logposterior}
\title{Evaluate mixture example posterior density}
\usage{
mixtureexample_logposterior(x)
}
\arguments{
\item{x}{evaluation points}
}
\value{
density values
}
\description{
Evaluate mixture example posterior density
}
