% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mixtureexample_logprior}
\alias{mixtureexample_logprior}
\title{Evaluate mixture example prior density}
\usage{
mixtureexample_logprior(x)
}
\arguments{
\item{x}{evaluation points}
}
\value{
density values
}
\description{
Evaluate mixture example prior density
}
