% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mixtureexample_posterior_cov}
\alias{mixtureexample_posterior_cov}
\title{Compute posterior covariance of mixture example}
\usage{
mixtureexample_posterior_cov(component)
}
\arguments{
\item{component}{of mixture}
}
\value{
posterior covariance
}
\description{
Compute posterior covariance of mixture example
}
