% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mixtureexample_posterior_mean}
\alias{mixtureexample_posterior_mean}
\title{Compute posterior mean of mixture example}
\usage{
mixtureexample_posterior_mean(component)
}
\arguments{
\item{component}{of mixture}
}
\value{
posterior mean
}
\description{
Compute posterior mean of mixture example
}
