% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mixtureexample_sampleprior}
\alias{mixtureexample_sampleprior}
\title{Sample from mixture example prior distribution}
\usage{
mixtureexample_sampleprior(n)
}
\arguments{
\item{n}{number of samples}
}
\value{
samples
}
\description{
Sample from mixture example prior distribution
}
