% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mixturemodel_gibbsvelocity}
\alias{mixturemodel_gibbsvelocity}
\title{Compute Gibbs flow for mixture model}
\usage{
mixturemodel_gibbsvelocity(
  time,
  xpoint,
  exponent,
  observations,
  mixturelogweights
)
}
\arguments{
\item{time}{time}

\item{xpoint}{particle position}

\item{exponent}{exponent of tempering schedule}

\item{observations}{dataset}
}
\value{
gibbs_velocity gibbs velocity field
}
\description{
Compute Gibbs flow for mixture model
}
