% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mixturemodel_gradloglikelihood}
\alias{mixturemodel_gradloglikelihood}
\title{Evaluate gradient of mixture model loglikelihood function}
\usage{
mixturemodel_gradloglikelihood(x, observations, mixturelogweights)
}
\arguments{
\item{x}{evaluation points}
}
\value{
gradient values
}
\description{
Evaluate gradient of mixture model loglikelihood function
}
