% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mixturemodel_gradlogprior}
\alias{mixturemodel_gradlogprior}
\title{Evaluate gradient of mixture model prior density}
\usage{
mixturemodel_gradlogprior(x)
}
\arguments{
\item{x}{evaluation points}
}
\value{
gradient values
}
\description{
Evaluate gradient of mixture model prior density
}
