% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mixturemodel_loglikelihood}
\alias{mixturemodel_loglikelihood}
\title{Evaluate mixture model loglikelihood function}
\usage{
mixturemodel_loglikelihood(x, observations, mixturelogweights)
}
\arguments{
\item{x}{evaluation points}
}
\value{
density values
}
\description{
Evaluate mixture model loglikelihood function
}
