% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mixturemodel_logprior}
\alias{mixturemodel_logprior}
\title{Evaluate mixture model prior density}
\usage{
mixturemodel_logprior(x)
}
\arguments{
\item{x}{evaluation points}
}
\value{
density values
}
\description{
Evaluate mixture model prior density
}
