% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mixturemodel_partialderivative_loglikelihood}
\alias{mixturemodel_partialderivative_loglikelihood}
\title{Evaluate partial derivative of mixture model likelihood function}
\usage{
mixturemodel_partialderivative_loglikelihood(
  dim,
  x,
  observations,
  mixturelogweights
)
}
\arguments{
\item{dim}{dimension along which partial derivative is taken (dim = 0 is the first coordinate)}

\item{x}{evaluation point}
}
\value{
partial derivative value
}
\description{
Evaluate partial derivative of mixture model likelihood function
}
