% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mixturemodel_sampleprior}
\alias{mixturemodel_sampleprior}
\title{Sample from mixture model prior distribution}
\usage{
mixturemodel_sampleprior(n)
}
\arguments{
\item{n}{number of samples}
}
\value{
samples
}
\description{
Sample from mixture model prior distribution
}
