% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvn_cholesky_factor}
\alias{mvn_cholesky_factor}
\title{Compute Cholesky factor}
\usage{
mvn_cholesky_factor(sigma)
}
\arguments{
\item{sigma}{covariance matrix}

\item{mu}{mean vector}
}
\value{
density values
}
\description{
Compute Cholesky factor
}
