% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvnpdf}
\alias{mvnpdf}
\title{Evaluate multivariate Gaussian density}
\usage{
mvnpdf(x, mu, sigma)
}
\arguments{
\item{x}{evaluation points}

\item{mu}{mean vector}

\item{sigma}{covariance matrix}
}
\value{
density values
}
\description{
Evaluate multivariate Gaussian density
}
