% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvnrnd}
\alias{mvnrnd}
\title{Simulate from multivariate Gaussian distribution}
\usage{
mvnrnd(n, mu, sigma)
}
\arguments{
\item{n}{number of samples}

\item{mu}{mean vector}

\item{sigma}{covariance matrix}
}
\value{
samples
}
\description{
Simulate from multivariate Gaussian distribution
}
