% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvnrnd_chol}
\alias{mvnrnd_chol}
\title{Simulate from multivariate Gaussian distribution (with pre-computed Cholesky factor)}
\usage{
mvnrnd_chol(n, mu, chol)
}
\arguments{
\item{n}{number of samples}

\item{mu}{mean vector}

\item{chol}{Cholesky factor of covariance matrix}
}
\value{
samples
}
\description{
Simulate from multivariate Gaussian distribution (with pre-computed Cholesky factor)
}
