% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gibbsflow_sis.R
\name{run_gibbsflow_sis}
\alias{run_gibbsflow_sis}
\title{Run Gibbs flow sequential importance sampler}
\usage{
run_gibbsflow_sis(
  prior,
  likelihood,
  nparticles,
  timegrid,
  lambda,
  derivative_lambda,
  compute_gibbsflow,
  gibbsvelocity = NULL
)
}
\arguments{
\item{prior}{list with keys: 
\code{logdensity} evaluates log prior density, 
\code{gradlogdensity} returns its gradient, 
\code{rinit} samples from the prior distribution}

\item{likelihood}{list with keys: 
\code{logdensity} samples from proposal, 
\code{gradlogdensity} returns its gradient}

\item{nparticles}{number of particles}

\item{timegrid}{vector describing numerical integration times}

\item{lambda}{vector describing tempering schedule}

\item{derivative_lambda}{time derivative of tempering schedule}

\item{compute_gibbsflow}{function computing Gibbs flow}
}
\value{
list with keys: 
\code{xtrajectory} trajectories, 
\code{xparticles} particles at terminal time,
\code{ess} effective sample size, 
\code{log_normconst} log normalizing constant,
}
\description{
Run Gibbs flow sequential importance sampler
}
\seealso{
\code{\link{run_gibbsflow_smc}} if resampling is desired
}
