% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gibbsflow_sisr.R
\name{run_gibbsflow_sisr}
\alias{run_gibbsflow_sisr}
\title{Run Gibbs flow importance sampler with resampling}
\usage{
run_gibbsflow_sisr(
  prior,
  likelihood,
  nparticles,
  timegrid,
  lambda,
  derivative_lambda,
  compute_gibbsflow
)
}
\arguments{
\item{prior}{list with keys: 
\code{logdensity} evaluates log prior density, 
\code{gradlogdensity} returns its gradient, 
\code{rinit} samples from the prior distribution}

\item{likelihood}{list with keys: 
\code{logdensity} samples from proposal, 
\code{gradlogdensity} returns its gradient}

\item{nparticles}{number of particles}

\item{timegrid}{vector describing numerical integration times}

\item{lambda}{vector describing tempering schedule}

\item{derivative_lambda}{time derivative of tempering schedule}

\item{compute_gibbsflow}{function computing Gibbs flow}
}
\value{
list with keys: 
\code{xtrajectory} trajectories, 
\code{xparticles} particles at terminal time,
\code{ess} effective sample size, 
\code{log_normconst} log normalizing constant,
}
\description{
Run Gibbs flow importance sampler with resampling
}
\seealso{
\code{\link{run_gibbsflow_sis}} if no resampling is desired
}
