% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{systematic_resampling}
\alias{systematic_resampling}
\title{Perform systematic resampling}
\usage{
systematic_resampling(normweights, nparticles, uniform)
}
\arguments{
\item{normweights}{normalized particle weights}

\item{nparticles}{number of particles}

\item{uniform}{uniform variable between 0 to 1 / nparticles}
}
\value{
\code{ancestors} ancestor indices
}
\description{
Perform systematic resampling
}
