% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{varcomp_artificial_logprior}
\alias{varcomp_artificial_logprior}
\title{Evaluate variance components model artificial prior density on simulated dataset}
\usage{
varcomp_artificial_logprior(x, dimension)
}
\arguments{
\item{x}{evaluation points}
}
\value{
density values
}
\description{
Evaluate variance components model artificial prior density on simulated dataset
}
