% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{varcomp_gibbsvelocity}
\alias{varcomp_gibbsvelocity}
\title{Compute Gibbs velocity for variance components model on simulated dataset}
\usage{
varcomp_gibbsvelocity(time, xparticles, exponent, dataset_rowsums, dimension)
}
\arguments{
\item{time}{time}

\item{xparticles}{particle positions}

\item{exponent}{exponent of tempering schedule}
}
\value{
gibbs_velocity gibbs velocity field
}
\description{
Compute Gibbs velocity for variance components model on simulated dataset
}
