% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{varcomp_gradloglikelihood}
\alias{varcomp_gradloglikelihood}
\title{Evaluate gradient of variance components loglikelihood on simulated dataset}
\usage{
varcomp_gradloglikelihood(x, dataset_rowsums, dimension, nrepetitions)
}
\arguments{
\item{x}{evaluation points}
}
\value{
gradient values
}
\description{
Evaluate gradient of variance components loglikelihood on simulated dataset
}
