% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{varcomp_gradlogprior_artificial}
\alias{varcomp_gradlogprior_artificial}
\title{Evaluate gradient of variance components artificial prior density on simulated dataset}
\usage{
varcomp_gradlogprior_artificial(x, dimension)
}
\arguments{
\item{x}{evaluation points}
}
\value{
gradient values
}
\description{
Evaluate gradient of variance components artificial prior density on simulated dataset
}
