% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{varcomp_loglikelihood}
\alias{varcomp_loglikelihood}
\title{Evaluate variance components model loglikelihood function on simulated dataset}
\usage{
varcomp_loglikelihood(x, dataset)
}
\arguments{
\item{x}{evaluation points}
}
\value{
density values
}
\description{
Evaluate variance components model loglikelihood function on simulated dataset
}
