% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{varcomp_sample_artificial_prior}
\alias{varcomp_sample_artificial_prior}
\title{Sample from variance components model artificial prior distribution}
\usage{
varcomp_sample_artificial_prior(N, dimension)
}
\arguments{
\item{n}{number of samples}
}
\value{
samples
}
\description{
Sample from variance components model artificial prior distribution
}
